Break enhance tracing code out of performance controller
